requirements.txt
streamlit-browser-engine